Fix block definition & misc changes

+++ shaders/lib/volumetricClouds.glsl
@@ -220,7 +220,7 @@ vec3 getCloudLighting(
 
 vec4 raymarchCloud(
 	int LayerIndex,
-	float samples,
+	int samples,
 	vec3 rayPosition,
 	vec3 rayDirection,
 	float dither,
@@ -319,7 +319,7 @@ vec4 raymarchCloud(
 			skylightOcclusion = mix(1.0, skylightOcclusion, distanceFade);
 		#endif
 
-		for(int i = 0; i < int(samples); i++) {
+		for(int i = 0; i < samples; i++) {
 
 			// check if the ray staring position is going farther than the reference distance, if yes, dont begin marching. this is to check for intersections with the world.
 			#ifndef VL_CLOUDS_DEFERRED
@@ -442,8 +442,8 @@ vec4 GetVolumetricClouds(
  		NormPlayerPos.y += 0.03 * heightRelativeToClouds;
  	#endif
 
-	int maxSamples = 15;
-	int minSamples = 10;
+	float maxSamples = 15.0;
+	float minSamples = 10.0;
 	int samples = int(clamp(maxSamples / sqrt(exp2(NormPlayerPos.y)),0.0, minSamples));
    
    	///------- setup the ray
